+++ Cmdlets/Public/Get-TeamViewerAuditEvent.ps1
@@ -1,4 +1,4 @@
-﻿function Get-TeamViewerEventLog {
+﻿function Get-TeamViewerAuditEvent {
     [CmdletBinding(DefaultParameterSetName = 'RelativeDates')]
 
     [OutputType('TeamViewerPS.AuditEvent')]